Fix Several Compile Warnings
