+++ hooks
@@ -665,14 +665,14 @@ ItemUseCB_RareCandy 081262AC 2
 
 DrawMoveInfoUIMarkers 08133404 0
 TMCase_MoveCursor_UpdatePrintedDescription 08131F0C 1
-CreateTMSprite 0813368c 1
+NullFunc 0813368c 1
 InitWindowTemplatesAndPals 08133244 1
 PrintStringTMCaseOnWindow3 081333c4 0
-SetTMSpriteAnim 08133714 2
-TintTMSpriteByType 08133730 1
-UpdateTMSpritePosition 08133790 2
-InitSelectedTMSpriteData 081337e4 2
-SpriteCB_MoveTMSpriteInCase 08133810 1
+NullFunc 08133714 2
+NullFunc 08133730 1
+NullFunc 08133790 2
+NullFunc 081337e4 2
+NullFunc 08133810 1
 
 ##CreateTMCaseSpriteHook 8131B94 0
 ##TMCaseMonIconPalChangeHook 8131E5C 0
