+++ include/start_menu.h
@@ -11,6 +11,5 @@ void __attribute__((long_call)) DestroySafariZoneStatsWindow(void);
 void __attribute__((long_call)) AppendToList(u8* list, u8* pos, u8 newEntry);
 
 void DrawTime(void);
-static bool8 ReloadStartMenu(void);
 
 #endif // GUARD_START_MENU_H
