+++ src/item.c
@@ -1831,7 +1831,7 @@ void ForceRedrawItemDescription(void)
 }
 void GetBargainItem(void) {
 	u16 items[] = { ITEM_CRACKED_POT, ITEM_PROTECTOR, ITEM_RAZOR_CLAW, ITEM_REAPER_CLOTH, ITEM_METAL_COAT, ITEM_QUICK_CLAW, ITEM_BINDING_BAND, ITEM_BLACK_SLUDGE, ITEM_CHARCOAL, ITEM_CHIPPED_POT, ITEM_DRAGON_FANG, ITEM_FOCUS_BAND, ITEM_METRONOME, ITEM_MIRACLE_SEED, ITEM_POISON_BARB, ITEM_PROTECTIVE_PADS, ITEM_RING_TARGET, ITEM_SPELL_TAG, ITEM_TWISTED_SPOON, ITEM_BLACK_BELT, ITEM_BLACK_GLASSES, ITEM_MAGNET, ITEM_NEVER_MELT_ICE, ITEM_PIXIE_PLATE, ITEM_SHARP_BEAK, ITEM_SILK_SCARF, ITEM_SILVER_POWDER };
-	for (int i = 0; i < ARRAY_COUNT(items); i++) {
+	for (u32 i = 0; i < ARRAY_COUNT(items); i++) {
 		if (Var8004 == items[i]) {
 			return;
 		}
@@ -1933,7 +1933,7 @@ void GetBargainItem(void) {
 }
 void GetSellItem(void) {
 	u16 items[] = { ITEM_TINY_MUSHROOM, ITEM_PEARL, ITEM_BIG_MUSHROOM, ITEM_SOFT_SAND, ITEM_RARE_BONE, ITEM_BIG_PEARL, ITEM_STAR_PIECE, ITEM_BALM_MUSHROOM, ITEM_PEARL_STRING, ITEM_COMET_SHARD };
-	for (int i = 0; i < ARRAY_COUNT(items); i++) {
+	for (u32 i = 0; i < ARRAY_COUNT(items); i++) {
 		if (Var8004 == items[i]) {
 			return;
 		}
